# internal/user/domain/DTO/user.go
package dto

type UserRequestDTO struct {
	FirstName string `json:"first_name"`
	LastName  string `json:"last_name"`
	Email     string `json:"email"`
	Password  string `json:"password"`
}

type UserResponseDTO struct {
	ID        string `json:"id"`
	FirstName string `json:"first_name"`
	LastName  string `json:"last_name"`
	Email     string `json:"email"`
	Role      string `json:"role"`
	CreateAt  string `json:"create_at"`
	UpdateAt  string `json:"update_at"`
}

type PatchRequestDTO struct {
	FirstName string `json:"first_name"`
	LastName  string `json:"last_name"`
	Email     string `json:"email"`
}

type ErrorResponse struct {
	Code  int    `json:"code"`
	Error string `json:"error"`
}